+++ joint_development/src/main/java/com/example/joint_development/repository/UserDetailMapper.java
@@ -3,6 +3,7 @@ package com.example.joint_development.repository;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import com.example.joint_development.domain.JointUser;
 import com.example.joint_development.domain.LangDetail;
 import com.example.joint_development.domain.UserDetail;
 
@@ -17,5 +18,10 @@ public interface UserDetailMapper {
 
     /** ユーザー取得（1件） */
     public UserDetail findOne(@Param("userId")Integer userId);
+    
+    /** ログインユーザー取得 */
+    public UserDetail findLoginUser(String email, String password);
+
+   
 
 }
