+++ joint_development/src/main/java/com/example/joint_development/service/JointUserService.java
@@ -19,11 +19,13 @@ public class JointUserService {
     public void signup(JointUser user){
         mapper.insertOne(user);
     }
-
+     
     /** ログインユーザー取得 */
     public JointUser getLoginUser(String email, String password){
         return mapper.findLoginUser(email, password);
     }
+
+   
 
     /** ユーザー取得 */
     public List<JointUser> getUsers(JointUser user){
