+++ joint_development/src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.sql.init.encoding=UTF-8
 
 
 #Spring
-server.servlet.context-path= /jointDevelopmnet
+server.servlet.context-path= /jointDevelopment
 
 #MyBatis
 mybatis.mapper-locations=classpath*:/mapper/jdbc/*.xml
